Support connecting external protocol adapters
The example AMQP Messaging Network and Device Registry can now be
configured to allow protocol adapters that run outside of the cluster to
connect to their secure endpoints.
This is helpful e.g. for connecting edge nodes which run a Hono protocol
adapter to connect to a Hono instance running in a back end.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>