Make tag name of Hono images configurable
The chart has supported configuration of the container image name for
all components already. However, the tag name has been part of the image
name so far. In order to allow setting the version of all Hono
components in one go, the honoImagesTag property has been introduced
which defaults to the chart's appVersion. The property can be used for
example to use not yet released (milestone) images for experimenting
purposes. All of Hono's components also support a specific imageTag
property to set the image's tag name per component.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>