Consolidate dictionaries passed into helper templates
The helper templates expect certain keys and values to be contained in
the scopes being passed in. The $args variable defined in the deployment
templates has been extended to also include the component specific
configuration properties from the values.yaml file. This allows the
default $args variable to be passed in to most helper templates as-is
which reduces overall complexity of the deployment templates.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>