Always pull images that have a SNAPSHOT tag
The default image pull policy is "IfNotPresent" which pulls images once
only, if they are not (yet) in the local cache. However, when testing
SNAPSHOT images being pulled in from a (private) registry, it makes more
sense to change that policy to "Always" in order to use the latest
available container image. A SNAPSHOT image in this context is an image
whose tag contains the term "SNAPSHOT".

Note that kubernetes applies the same logic for container images having
a "latest" tag.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>