Use back end to centrally manage Jaeger sampling strategy
The Jaeger Collector can serve sampling strategies for individual
services to Agents from a central configuration file.

The chart now uses this approach to define a default "sample everything"
strategy and distribute it to all Hono components' side car Agent
containers.

The sampling strategies can be configured in a custom file as well which
can be mounted as an extra secret into the Jaeger all-in-one container.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>