[#222] Support Kafka based messaging in Hono.
This adds the support for the Kafka based messaging to the Hono chart.
The property `messagingNetworkType` is used to select which type of
configuration is added to the adapters and services (either `amqp` or
`kafka`). The property `kafkaMessagingClusterExample.enabled` indicates
whether the example Kafka cluster should be deployed and used.

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>