[#222] Use TLS for connecting to Kafka.
New certificates are created for the example Kafka broker. The broker
is configured to expect TLS encrypted connections from clients. The
truststore is added to all services and adapters that need to connect
to Kafka. This is currently a "jks" file because in Kafka the support for
the "pem" format was added in version 2.7 and Hono currently uses the
Kafka clients in version 2.6.

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>