Replace usage of pem truststore by JKS.
Kafka (in the version currently used by Hono) does not support PEM files.
Instead of providing the same certificates in PEM as well as in JKS format
to Hono's components, this replaces the usage of the file `trusted-certs.pem`
with the file `trustStore.jks`. The JKS file requires a password which is
added to the configuration as well.

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>