Fix cloud2edge's post-install running before its deps are available
post-install runs when the chart is deployed, not when it's actually
ready, so it's entirely possible for the post-install hook to try to
access services that have not even started yet, resulting in a
deployment failure. Instead, wait for the needed services to be ready,
following the same pattern as used by the nginx-deployment template.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>